+++ tests/Support/Stub/DefaultProvider.php
@@ -6,6 +6,7 @@ namespace UIAwesome\Html\Svg\Tests\Support\Stub;
 
 use UIAwesome\Html\Core\Base\BaseTag;
 use UIAwesome\Html\Core\Provider\DefaultsProviderInterface;
+use UIAwesome\Html\Svg\Base\BaseSvgTag;
 
 /**
  * Provides default configuration values for tag instances in HTML helper and UI component testing.
@@ -32,9 +33,14 @@ final class DefaultProvider implements DefaultsProviderInterface
      */
     public function getDefaults(BaseTag $tag): array
     {
-        return [
-            'class' => 'default-class',
-            'title' => 'default-title',
-        ];
+        return match (true) {
+            $tag instanceof BaseSvgTag => [
+                'class' => 'default-class',
+            ],
+            default => [
+                'class' => 'default-class',
+                'title' => 'default-title',
+            ],
+        };
     }
 }