+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 0.3.0 Under development
 
-- Enh #4: Refactor codebase to improve performance (@terabytesoftw)
-- Enh #5: Add `SvgBlock` enum for HTML SVG tag representation and implement `BlockInterface` class.
+- Enh #5: Refactor codebase to improve performance (@terabytesoftw)
+- Enh #6: Add `SvgBlock` enum for HTML SVG tag representation and implement `BlockInterface` class (@terabytesoftw)
+- Enh #7: Add `BaseSvgTag` and `Defs` classes for SVG block-level elements and related tests (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024
 
@@ -11,4 +12,4 @@
 
 ## 0.1.0 March 22, 2024
 
-- Enh #1: Initial release.
+- Enh #1: Initial release (@terabytesoftw)